Rewrite ACIs completely to support hostcategory and usercategory
When hostcategory or usercategory are set to 'all' on the rule,
content of the rule, including the profile data becomes available
to all authenticated users.

To support this, we should use a target filter that explicitly matches
those entries that have hostcategory/usercategory set to 'all'.

In addition, we need to make sure we are allowing to read/search/compare
entries which match our ACIs or otherwise host principal will not be
able to look up such entries. Thus, I'm also adding a wild card
targetattr in the ACI.

Finally, a CoS definition should be done as a separate LDAP sub-entry.
If we don't do it this way, then cn=rules,cn=desktop-profile,$SUFFIX
will not be visible in LDAP which is not only confusing but also makes
ACIs not working properly.